backend/pyproject.toml
[tool.poetry]
name = "ai-todo-backend"
version = "0.1.0"
description = "Backend for AI Todo Chatbot"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlmodel = "^0.0.14"
openai-agents-sdk = "^0.1.0"
mcp = "^0.1.0"
better-auth = "^0.1.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
openai = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
